+++ lib/presentation/splash/provider/splash_contract.dart
