+++ lib/presentation/splash/view/splash_page.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
-import 'package:picple/presentation/splash/controller/splash_contract.dart';
-import 'package:picple/presentation/splash/controller/splash_notifier.dart';
 
 import '../../theme/picple_colors.dart';
+import '../provider/splash_contract.dart';
+import '../provider/splash_notifier.dart';
 
 class SplashPage extends ConsumerWidget {
   const SplashPage({super.key});
